hivera/requirements.txt
telethon==1.34.0
aiohttp==3.8.5
python-dotenv==1.0.0
colorama==0.4.6
